[parameters.functional] change behavior for scalar parameters
WARNING: this could break your code! Before this commit, if you had a
parameter of type ('key': 1) (or any number of keys, but scalar ones), the
variable of the parameter functional was 'key'. If you had a parameter of
type ('key': 2), the variables were 'key[0]' and 'key[1]'. In order to
allow for more generic programs, this commit changes the variable in the
scalar case to 'key[0]'.